third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-contain/container-queries/font-relative-units-dynamic-expected.txt

This is a testharness.js-based test.
[PASS] em units respond to changes
[PASS] rem units respond to changes
[PASS] ex units respond to changes
[PASS] rex units respond to changes
[PASS] ch units respond to changes
[FAIL] cap units respond to changes
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
[PASS] rch units respond to changes
[PASS] lh units respond to changes
[PASS] rlh units respond to changes
[PASS] ic units respond to changes
[PASS] ric units respond to changes
[FAIL] rcap units respond to changes
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
